requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
python-dotenv==0.15.0
